src/api/requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
azure-identity
torch
torchvision
lightgbm
pandas
numpy
scikit-learn
python-multipart
langchain-community
langchain-text-splitters
langchain-huggingface
azure-search-documents
azure-core
sentence-transformers
pypdf
langchain-openai
langchain